backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.6
opencv-python==4.8.1.78
tensorflow==2.13.0
keras==2.13.1
# dlib==19.24.2  # Optional - can cause build issues on some systems
mediapipe==0.10.21
# pymongo==4.5.0  # Optional - only needed for MongoDB
# mysql-connector-python==8.1.0  # Optional - conflicts with mediapipe protobuf
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-dotenv==1.0.0
Pillow==10.0.0
requests==2.31.0
python-socketio==5.8.0
eventlet==0.33.3
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
